requirements.txt
numpy==1.19.1
face_recognition==1.3.0
matplotlib==3.2.2
Pillow==7.2.0
